+++ rizza/task_manager.py
@@ -4,6 +4,7 @@ import json
 from collections import deque
 from multiprocessing.pool import ThreadPool as Pool
 import attr
+from logzero import logger
 from rizza.entity_tester import EntityTestTask
 from rizza.helpers.misc import json_serial
 # joblib (run on multiple cores)
@@ -50,16 +51,14 @@ class TaskManager(object):
                 outfile.write("\n")
 
     @staticmethod
-    def log_tests(path=None, tests=None, mock=False):
-        """Run and log the tests passed in."""
-        with open(path, "w") as log:
-            print("Writing to log file: {0}".format(path))
-            for test in tests:
-                log.write("{0}~{1}~{2}\n".format(
-                    json.dumps(attr.asdict(test), default=json_serial),
-                    json.dumps(test.execute(mock), default=json_serial),
-                    json.dumps(attr.asdict(test), default=json_serial)
-                ))
+    def run_tests(tests=None, mock=False):
+        """Run the tests passed in."""
+        for test in tests:
+            logger.info("{0}~{1}~{2}\n".format(
+                json.dumps(attr.asdict(test), default=json_serial),
+                json.dumps(test.execute(mock), default=json_serial),
+                json.dumps(attr.asdict(test), default=json_serial)
+            ))
 
     @coroutine
     def run_tasks(self, limit=None, threads=1):